Reversing auto-partitioning work
The solution ended up being significantly more complex than I would
have liked. In most cases it's probably relatively easier to do
any necessary partitioning when actually defining the batch jobs. In
that context the merging strategy should also be clear (merging
partitioned batch requests in a general way is complex and will
necessarily be imperfect unless we impose minimum attr constraints on
the request).

Am leaving the WIP soln. in the repo history in case it (or part of it)
ever becomes useful in the future.

C'est la vie.