[#77-#84] BREAKING: new features, tests, etc. (@ricardojmendez)
* Renamed environment sample, more useful as a script
* Global secondary index tests for create-table
* Local secondary index tests

* BREAKING CHANGE: update-table now receives options in a map [1]
This commit only changes throughput so that it's passed as an option in
a map, as discussed on #78. It's still required, which it shouldn't be
since it's not the only option to change.

* We no longer need to do stepping when updating throughput
See discussion on #78

* Elaborating on table-status-watch comment
* Tests for decreasing throughput
* Disregarding throughput updates that are the same as the current values
Found that updateTable hangs if we pass the same throughput values as
the table currently has (see #79).  Refactored throughput parameter
validation to its own function, since we'll need to validate the
parameters for the other operations as well.

* Expanded tests.
* Updating encore and AWS SDK version, all tests passing
Want to make sure we're running against the latest version, to avoid
behavior like #79 being a known (and possibly solved) issue.

* Upgraded expectations, re-wrote tests that used expect-let
* Sending empty map to update-table should have no effect
See 8574687, this test should have gone in with that commit.

* Global secondary index creation on update-table, see #80
* update-table can now change throughput on :gsindexes (#80)
Also, new function: index-status-watch

* update-table can now delete :gsindexes (#80)
* Support for specifying index on scan (#82)
* Supporting projection and expr-attr-names on scan (#82)
* Added support for conditional deletes (#84)
* Added missing tests for expressions on update-item-request
* BREAKING CHANGE: update-map is now another value on update-item's params
Updated tests, fixed deprecation warning. See #83.

* Expanded docstring for update-table (#80)