Added --no-kill option to run-tests
I suspect that the root cause of #104, where tests on travis fail after
the first run-through, is that it's trapping the exit of the first test
round and killing DynamoDBLocal.  That doesn't happen on OS X, but seems
to happen on whatever distribution the container is running (Debian?).

This option should only be used if you want DynamoDBLocal to remain
running. In a containerized environment, that's a problem that takes
care of itself.