Add ability to manage TTL on tables
`update-ttl` and `describe-ttl` fns

I noticed that tests were failing on a fresh checkout and my alterations
meant they were failing for different reasons so I felt obliged to look
into it. I believe I've managed to keep all the tests relevant and to
the spirit with which they were intended. There are some differences
between the local and non-local testing systems which, again, I've tried
to cater for.