Treat merging of items differently from the rest of a result
This fixes an issue where `batch-get-item` results weren't merged correctly and we'd always end up with the final page of results.

It changes the `query` and `scan` functions to temporarily structure their results in the same way as `batch-get-item` (where the list of items is against a particular table name). These are then merged and the table name part of the result is removed.

I've used `update-in` with a single item vector because using `update` fails the tests for Clojure 1.5 and 1.6!