Avoid request 'stitching' when span-reqs is not specified
Until now, the default 'span-reqs' configuration stitched up to 5
requests together if last-prim-kvs implied it was necessary. This caused
confusion because it's not getting all results but it's also not
limited to just one page. The limit parameter also then became
confusing, because the number of items returned by a request would be
somewhere between '5' and '5 x limit'.

After this commit the various 'get items' functions will return only one
page by default and if subsequent pages are needed then span-reqs must
be used.

Closes #74